// frontend/src/app/api/models/rag-overview.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CollectionInfo } from '../models/collection-info';
import { EmbedderInfo } from '../models/embedder-info';
import { JobStatus } from '../models/job-status';
import { SourceInfo } from '../models/source-info';
import { StoreInfo } from '../models/store-info';
import { WriteAccess } from '../models/write-access';
export interface RagOverview {
  asOf: string;
  collections: Array<CollectionInfo>;
  embedder: EmbedderInfo;
  job?: (JobStatus | null);
  maxChars: number;
  sources: Array<SourceInfo>;
  store: StoreInfo;
  writes: WriteAccess;
}